requirements.txt
streamlit
smartapi-python
python-dotenv
pyotp
requests
